refactor: economy phase-2 slice 13 — worker action cost scenario tables (#3856)
Hoist recruit/train affordability and deduction rows into
worker_action_cost_scenarios.dart; runner uses table loops.
