+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -49,6 +49,7 @@ export 'src/purchased_tile_riches_scenarios.dart';
 export 'src/purchased_tile_riches_test_support.dart';
 export 'src/treasury_bid_budget_test_support.dart';
 export 'src/treasury_bid_budget_scenarios.dart';
+export 'src/worker_action_cost_scenarios.dart';
 export 'src/trade_order_admission_scenarios.dart';
 export 'src/trade_order_factory.dart';
 export 'src/trade_order_suggester_scenarios.dart';
